// frontend/src/app/model/travel-insurance.ts
export class TravelInsurance {
    id: string;
    name: string;
    price: string
    coverageType: string;
    coverageLimit: string;
    deductible: string;
    coveragePeriod: string;
    description: string;
    restriction: string;
    additionalInfo: string;
    ageLimit: string;
    sportCoverage: string;
    preExistingConditionCoverage: string;
    regionCoverage: string;
    insertUserId: string;
    updateUserId: string;
    insertDateTime: Date;
    updateDateTime: Date;
    status: number;
    sortOrder: number;
}